Push unsafety to the Rust std library
By making use of the built in methods copy_from_slice (stable since
1.9.0) and copy_within (stable since 1.37.0) we avoid having any
  unsafety ourselves.
This comes at the cost of potentially panicking if our assumptions don't
hold.

In addition I've tried to document the logic behind replace_slice a bit
more and simplified the case where the source and destination slice are
of equal length.